requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-groq
langchain-chroma
langchain-classic
langchain-huggingface
chromadb
fastapi
uvicorn
streamlit
python-dotenv
pypdf
python-multipart
requests
sentence-transformers
